+++ src/constants/index.ts
@@ -2,5 +2,4 @@ export * from './networks'
 export * from './others'
 export * from './endpoints'
 export * from './addresses'
-export * from './dataviewContents'
 export * from './enums'
